requirements.txt
appdirs==1.4.4
asgiref==3.4.1
black==21.7b0
cairocffi==1.0.0
CairoSVG==2.5.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.2.5
django-appconf==1.0.4
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-imagekit==4.0.2
django-js-asset==1.2.2
djangorestframework==3.12.4
html5lib==1.1
idna==3.2
mypy-extensions==0.4.3
pathspec==0.9.0
pilkit==2.0
Pillow==8.1.2
pycparser==2.20
pyphen==0.11.0
python-dotenv==0.18.0
pytz==2021.1
razorpay==1.2.0
regex==2021.7.6
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
tinycss2==1.1.0
tomli==1.1.0
urllib3==1.26.6
WeasyPrint==52.5
webencodings==0.5.1
